Use super class for generic operations
