+++ test/lib/test_to_code_mapping_test.rb
@@ -1,16 +1,10 @@
 require 'test_helper'
 require 'ttnt/test_to_code_mapping'
 
-class TestToCodeMappingTest < Minitest::Test
-  FIZZBUZZ_FIXTURE_DIR = File.expand_path('../../fixtures/repositories/fizzbuzz', __FILE__).freeze
-
+class TestToCodeMappingTest < TTNT::TestCase
   def setup
-    prepare_git_repository
+    super
     @test_to_code_mapping = TTNT::TestToCodeMapping.new(@repo, @repo.head.target_id)
-  end
-
-  def teardown
-    FileUtils.remove_entry_secure(@tmpdir)
   end
 
   def test_append_from_coverage
@@ -37,18 +31,5 @@ class TestToCodeMappingTest < Minitest::Test
       'It should be empty for code which is not executed'
     assert_equal Set.new([test_file]),
       @test_to_code_mapping.get_tests(file: 'lib/fizzbuzz.rb', lineno: 2)
-  end
-
-  private
-
-  def prepare_git_repository
-    @tmpdir = Dir.mktmpdir('ttnt_repo')
-    FileUtils.cp_r(FIZZBUZZ_FIXTURE_DIR, @tmpdir)
-    @repodir = "#{@tmpdir}/fizzbuzz"
-    Dir.chdir(@repodir) do
-      FileUtils.rm('.git')
-      File.rename('.gitted', '.git') if File.exist?(".gitted")
-    end
-    @repo = Rugged::Repository.new(@repodir)
   end
 end
